fix: reolve circular dependency
