+++ src/hooks/routing/useRouterTrade.ts
@@ -9,7 +9,7 @@ import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import useDebounce from 'hooks/useDebounce'
 import useIsValidBlock from 'hooks/useIsValidBlock'
 import useIsWindowVisible from 'hooks/useIsWindowVisible'
-import { useStablecoinAmountFromFiatValue } from 'hooks/useUSDCPrice'
+import { useStablecoinAmountFromFiatValue } from 'hooks/useStablecoinAmountFromFiatValue'
 import ms from 'ms.macro'
 import { useMemo } from 'react'
 import { useGetQuoteQuery } from 'state/routing/slice'
