+++ src/hooks/useUSDCPrice.ts
@@ -1,20 +1,7 @@
 import { Currency, CurrencyAmount, Price, Token, TradeType } from '@uniswap/sdk-core'
-import { SupportedChainId } from 'constants/chains'
-import { DAI_OPTIMISM, USDC_ARBITRUM, USDC_MAINNET, USDC_POLYGON } from 'constants/tokens'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { useMemo, useRef } from 'react'
-import tryParseCurrencyAmount from 'utils/tryParseCurrencyAmount'
-
 import { useRouterTrade } from './routing/useRouterTrade'
-
-// Stablecoin amounts used when calculating spot price for a given currency.
-// The amount is large enough to filter low liquidity pairs.
-export const STABLECOIN_AMOUNT_OUT: { [chainId: number]: CurrencyAmount<Token> } = {
-  [SupportedChainId.MAINNET]: CurrencyAmount.fromRawAmount(USDC_MAINNET, 100_000e6),
-  [SupportedChainId.ARBITRUM_ONE]: CurrencyAmount.fromRawAmount(USDC_ARBITRUM, 10_000e6),
-  [SupportedChainId.OPTIMISM]: CurrencyAmount.fromRawAmount(DAI_OPTIMISM, 10_000e18),
-  [SupportedChainId.POLYGON]: CurrencyAmount.fromRawAmount(USDC_POLYGON, 10_000e6),
-}
+import { STABLECOIN_AMOUNT_OUT } from './useStablecoinAmountFromFiatValue'
 
 /**
  * Returns the price in USDC of the input currency
@@ -63,29 +50,4 @@ export function useUSDCValue(currencyAmount: CurrencyAmount<Currency> | undefine
       return null
     }
   }, [currencyAmount, price])
-}
-
-/**
- *
- * @param fiatValue string representation of a USD amount
- * @returns CurrencyAmount where currency is stablecoin on active chain
- */
-export function useStablecoinAmountFromFiatValue(fiatValue: string | null | undefined) {
-  const { chainId } = useActiveWeb3React()
-  const stablecoin = chainId ? STABLECOIN_AMOUNT_OUT[chainId]?.currency : undefined
-
-  return useMemo(() => {
-    if (fiatValue === null || fiatValue === undefined || !chainId || !stablecoin) {
-      return undefined
-    }
-
-    // trim for decimal precision when parsing
-    const parsedForDecimals = parseFloat(fiatValue).toFixed(stablecoin.decimals).toString()
-    try {
-      // parse USD string into CurrencyAmount based on stablecoin decimals
-      return tryParseCurrencyAmount(parsedForDecimals, stablecoin)
-    } catch (error) {
-      return undefined
-    }
-  }, [chainId, fiatValue, stablecoin])
 }
